refactor: Extract main view from MainActivity
